io: compatibility

+++ src/io/size/s1.c
@@ -17,10 +17,10 @@ main(int argc, char **argv)
       exit(EXIT_FAILURE);
     }
 
-  printf("size: %lld\nblock size: %ld\nblocks: %lld\n",
-         buf.st_size,
+  printf("size: %lld\nblock size: %lld\nblocks: %lld\n",
+         (long long int) buf.st_size,
          (long long int) buf.st_blksize,
-         (long int) buf.st_blocks);
+         (long long int) buf.st_blocks);
 
 
   exit(EXIT_SUCCESS);
